pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cns"
version = "0.0.1"
description = "Cognition-Native Store"
requires-python = ">=3.10"
dependencies = [
  "psycopg[binary]>=3.1",
  "pgvector>=0.2.5",
  "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "hypothesis>=6.112.0",
  "testcontainers[postgres]>=4.9.0",
  "ruff>=0.6.9",
  "black>=24.8.0",
  "mypy>=1.10.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = ["E", "F", "I"]

[tool.setuptools]
packages = ["cns_py"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_no_return = true
ignore_missing_imports = true
packages = ["cns_py"]
